backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mediapipe==0.10.14
ml-dtypes==0.4.0
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
openai==1.38.0
openai-whisper==20231117
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.4
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
sympy==1.13.1
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whisper==1.1.10
yt-dlp==2024.8.6
torch==2.4.0